Backend: Fixing a small bugs in migration

+++ server/database/migrations/2022_11_14_181841_modify_pickups_table.php
@@ -10,8 +10,8 @@ return new class extends Migration
     {
         Schema::table('pickups', function (Blueprint $table) {
             $table->boolean("live_location")->default(false)->after("canceled");
-            $table->integer("latitude")->default(-1)->after("live_location");        
-            $table->integer("longitude")->default(-1)->after("latitude");
+            $table->double("latitude")->default(-1)->after("live_location");        
+            $table->double("longitude")->default(-1)->after("latitude");
             
         });
     }
